# diplom/cashFlowManager/cfm-api/swagger-api-specification/api-specification.yaml
openapi: 3.0.0
info:
  title: Cash Flow Manager API
  version: 0.0.1
...
paths:
  /owner/{ownerId}/:
    get:
      tags:
        - owner
      summary: Получение данных пользователя
      operationId: getOwnerInfo
      parameters:
        - name: ownerId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Успешный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OwnerPersonalDataInfoResponse'
        '400':
          description: Некорректный запрос от клиента
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Неавторизованный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Запрашиваемый ресурс не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /owner/list/:
    get:
      tags:
        - owner
      summary: Получение списка всех пользователей
      operationId: getOwnersList
      responses:
        '200':
          description: Успешный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OwnersListResponse'
        '400':
          description: Некорректный запрос от клиента
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Неавторизованный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Запрашиваемый ресурс не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /owner/add/:
    post:
      tags:
        - owner
      summary: Добавление информации о владельце счета
      operationId: addOwner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddOwnerRequest'
      responses:
        '200':
          description: Успешное добавление информации о владельце счета
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddOwnerResponse'
        '400':
          description: Некорректный запрос от клиента
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Неавторизованный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Запрашиваемый ресурс не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /owner/update/{ownerId}/:
    post:
      tags:
        - owner
      summary: Изменение информации о владельце счета
      operationId: updateOwner
      parameters:
        - name: ownerId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOwnerRequest'
      responses:
        '200':
          description: Успешное изменение информации о владельце счета
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateOwnerResponse'
        '400':
          description: Некорректный запрос от клиента
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Неавторизованный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Запрашиваемый ресурс не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /owner/delete/{ownerId}/:
    post:
      tags:
        - owner
      summary: Удаление владельца счета
      operationId: deleteOwner
      parameters:
        - name: ownerId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Успешное удаление владельца счета
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteOwnerResponse'
        '400':
          description: Некорректный запрос от клиента
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Неавторизованный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Запрашиваемый ресурс не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /owner/deleted/list/:
    get:
      tags:
        - owner
      summary: Получение списка всех удаленных пользователей
      operationId: getDeletedUsersList
      responses:
        '200':
          description: Успешный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeletedOwnersListResponse'
        '400':
          description: Некорректный запрос от клиента
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Неавторизованный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Запрашиваемый ресурс не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    ErrorResponse:
      description: Ошибка от приложения
      type: object
      required:
        - status
        - error
        - message
        - exception
        - path
        - timestamp
      properties:
        status:
          description: Статус ошибки
          type: string
          example: "500"
          pattern: "^[45]0[0-9]{1}$"  # Ошибка сервера (5xx) или клиента (4xx)
          maxLength: 3  # Максимальная длина строки
        error:
          description: Описание ошибки
          type: string
          example: "Internal Server Error"
        message:
          description: Сообщение об ошибке
          type: string
          example: "Произошла ошибка при обработке запроса"
        exception:
          description: Имя исключения
          type: string
          example: "NullPointerException"
        path:
          description: Путь запроса, вызвавшего ошибку
          type: string
          example: "/api/v1/users/123"
        timestamp:
          description: Временная метка ошибки
          type: string
          format: date-time  # Формат даты-времени
          example: "2024-02-06T12:30:45Z"
      additionalProperties: false
    OwnerPersonalDataItem:
      description: "Данные о пользователе"
      type: "object"
      required:
        - id
        - ownerSurname
        - ownerName
        - email
        - phoneNumber
        - address
        - dateOfBirth
        - gender
        - status
        - createdAt
      properties:
        id:
          description: "Идентификатор персональных данных владельца"
          type: "string"
        ownerSurname:
          description: "Фамилия владельца"
          type: "string"
        ownerName:
          description: "Имя владельца"
          type: "string"
        ownerPatronymic:
          description: "Отчество владельца"
          type: "string"
        email:
          description: "Email владельца"
          type: "string"
        phoneNumber:
          description: "Номер телефона владельца"
          type: "string"
        address:
          description: "Адрес владельца"
          type: "string"
        dateOfBirth:
          description: "Дата рождения владельца"
          type: integer
          format: int64
        gender:
          description: "Пол владельца"
          type: "string"
        status:
          description: "Статус владельца счета"
          type: "string"
        createdAt:
          description: "Дата создания записи о владельце счета"
          type: integer
          format: int64
      additionalProperties: false
...
    OwnersListResponse:
      description: "Ответ со списком всех пользователей"
      type: "object"
      required:
        - ownersList
      properties:
        'accountsList':
          description: "Список всех пользователей"
          type: array
          items:
            $ref: '#/components/schemas/OwnerPersonalDataItem'
      additionalProperties: false
...
    OwnerPersonalDataInfoResponse:
      description: "Запрос на добавление счета"
      required:
        - owner
      type: "object"
      properties:
        'account':
          $ref: '#/components/schemas/OwnerPersonalDataItem'
...
    AddOwnerRequest:
      description: "Запрос на добавление клиента"
      type: "object"
      required:
        - owner
      properties:
        'owner':
          $ref: '#/components/schemas/OwnerPersonalDataItem'
    AddOwnerResponse:
      description: "Ответ на добавление клиента"
      type: "object"
      required:
        - ownerId
      properties:
        ownerId:
          description: Идентификатор только что добавленного клиента
          type: "string"
    UpdateOwnerRequest:
      description: "Запрос на изменения информации о клиенте"
      type: "object"
      required:
        - owner
      properties:
        'owner':
          $ref: '#/components/schemas/OwnerPersonalDataItem'
    UpdateOwnerResponse:
      description: "Ответ на изменения информации о клиенте"
      type: "object"
      required:
        - owner
      properties:
        'owner':
          $ref: '#/components/schemas/OwnerPersonalDataItem'
    DeleteOwnerResponse:
      description: "Запрос на удаление владельца счета"
      type: "object"
      required:
        - ownerId
      properties:
        ownerId:
          description: Идентификатор только что добавленного клиента
          type: "string"
    DeletedOwnersListResponse:
      description: "Ответ со списком всех удаленных клиентов"
      type: "object"
      required:
        - deletedOwnersList
      properties:
        'deletedOwnersList':
          description: "Список всех удаленных клиентов"
          type: array
          items:
            $ref: '#/components/schemas/OwnerPersonalDataItem'
...